0.2.0
getpid(2) returns a signed integer but the call is guaranteed to return a
positive integer. PIDs of 0 and -1 are special values (self and all PIDs).

getsockopt(SO_PEERCRED) does not document valid PIDs returned in the pid
field of struct ucred/sockpeercred. Presumably it has the same guarantees
as getpid(2) since the process is running on the same system.

Make the behaviour explicit by setting the UNIXREMOTEPID environment
variable only if the remote pid field is greater than 0.